Update from task 37f2e80c-c5c5-4dd8-907f-875a39631d94

+++ CMakeProject1/src/AISystem.cpp
@@ -289,7 +289,7 @@ namespace CmakeProject1 {
     }
 
     int AISystem::GetActiveBrainCount() const {
-        return aiBrains.size();
+        return static_cast<int>(aiBrains.size());
     }
 
 }
